Fix single test case line core segmentation fault
Error with old logic led to a segmentation fault.

Key started at 1 which would be our index iterator for our loop.

If inputMap.size() = 1, the loop would not start due to it being <
exclusive.

This would leave an empty string to be returned.

message.erase(message.size()-1) attempted to modify an empty string, causing the fault.

Solution; a while loop. Realistically, it does not matter what our upper
bound was for a loop because of our inner logic of breaking early
regardless.